requirements.txt
# Core
streamlit>=1.30.0
python-dotenv>=1.0.0

# Google Gemini API
google-generativeai>=0.5.0

# Data Processing
pandas>=2.1.0
numpy>=1.25.0

# Visualization
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Forecasting (Optional)
statsmodels>=0.14.0

# PDF Report Generation
fpdf2>=2.7.0

# Synthetic Data Generation
faker>=22.0.0

# Utilities
openpyxl>=3.1.0
